cleanup samples
